requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-experimental
langchain-community
langchain-tavily
playwright
e2b
python-dotenv
PyYAML
tenacity
fastapi
uvicorn
sse-starlette